+++ CHANGELOG.md
@@ -8,6 +8,8 @@ Types of changes: `Added`, `Changed`, `Deprecate`, `Removed`, `Fixed`, `Secruity
 
 ## [Unreleased]
 
+## v2.3.0 - 10.03.2024 - Improvements to the framework and a new volume distribution statistic
+
 ### Added
 - Added function to remove leading spaces on the beginning of each line
 - Added support for `in` condition in query builder